+++ app/src/main/java/com/example/location_feature/dependencyInjection/RepositoryModule.kt
@@ -1,0 +1,16 @@
+package com.example.location_feature.dependencyInjection
+import com.example.location_feature.network.FirebaseAuthRepositoryImpl
+
+import com.example.location_feature.domain.repository.AuthRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Binds
+    abstract fun bindAuthRepository(authRepository: FirebaseAuthRepositoryImpl): AuthRepository
+}
