--- app/src/main/java/com/example/location_feature/view/ui/Fragments/HomeFragment.kt
@@ -1,29 +1,0 @@
-package com.example.location_feature.view.ui.Fragments
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.location_feature.databinding.FragmentHomeMapaBinding
-
-class HomeFragment : Fragment() {
-
-    private var _binding: FragmentHomeMapaBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        // Inflate the layout for this fragment
-        _binding = FragmentHomeMapaBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-}
